Add RTP packetizer
This packetizer sends H.264 video stream using
RTP frames defined by RFC 6184 and RFC 4571.
For this, part of SDLStreamingMediaManager is separated
as SDLH264ByteStreamPacketizer.
Spec files for both SDLH264ByteStreamPacketizer and
SDLRTPH264Packetizer are added.

This commit also includes a temporal change
to test RTP stream output. To enable it, change "useRTP"
flag to YES in SDLStreamingMediaManager.m.

This is for discussion on proposal SDL-0048.